Groovy classes to Java
