+++ remoting/dolphin-remoting-shared/src/main/groovy/org/opendolphin/core/comm/Command.java
@@ -1,20 +1,6 @@
-/*
- * Copyright 2015-2016 Canoo Engineering AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.opendolphin.core.comm
-// todo dk: go through all subtypes and apply the new naming convention
+package org.opendolphin.core.comm;
+
+import org.codehaus.groovy.runtime.StringGroovyMethods;
 
 /**
  * Commands come in two flavors: *Command (active voice) and *Notification (passive voice).
@@ -26,9 +12,7 @@ package org.opendolphin.core.comm
  * They are only "DTOs" that are sent over the wire.
  * The receiving side is responsible for finding the appropriate action.
  */
-
 public class Command {
-
     public Command() {
 
     }
@@ -38,10 +22,11 @@ public class Command {
     }
 
     public static String idFor(Class commandClass) {
-        commandClass.getSimpleName() - "Command" - "Notification"
+        return StringGroovyMethods.minus(StringGroovyMethods.minus(commandClass.getSimpleName(), "Command"), "Notification");
     }
 
     public String toString() {
         return "Command: " + getId();
     }
+
 }
